Add userlist to fragment
