+++ app/build.gradle
@@ -73,4 +73,8 @@ dependencies {
     def exoPlayerVersion = "2.15.0"
     implementation "com.google.android.exoplayer:exoplayer-core:$exoPlayerVersion"
     implementation "com.google.android.exoplayer:exoplayer-ui:$exoPlayerVersion"
+
+    // Fragment
+    def fragmentVersion = "1.3.6"
+    implementation "androidx.fragment:fragment-ktx:$fragmentVersion"
 }
