+++ app/src/main/java/com/daniillyubaev/ourawesomeapp/MainActivity.kt
@@ -19,48 +19,4 @@ class MainActivity : AppCompatActivity(R.layout.activity_main) {
     companion object {
         val LOG_TAG = "MyAwesomeLogTag"
     }
-
-    private val viewModel: MainViewModel by viewModels()
-    private val viewBinding by viewBinding(ActivityMainBinding::bind)
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        Log.d(LOG_TAG, "onCreate()")
-        setContentView(R.layout.activity_main)
-        setupRecyclerView()
-        lifecycleScope.launch {
-            lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
-                viewModel.viewState.collect { viewState ->
-                    Log.d(LOG_TAG, "$viewState")
-                   renderViewState(viewState)
-                }
-            }
-        }
-    }
-
-    private fun renderViewState(viewState: MainViewModel.ViewState) {
-        when (viewState) {
-            is MainViewModel.ViewState.Loading -> {
-                viewBinding.usersRecyclerView.isVisible = false
-                viewBinding.progressBar.isVisible = true
-            }
-            is MainViewModel.ViewState.Data -> {
-                viewBinding.usersRecyclerView.isVisible = true
-                (viewBinding.usersRecyclerView.adapter as UserAdapter).apply{
-                    userList = viewState.userList
-                    notifyDataSetChanged()
-                }
-                viewBinding.progressBar.isVisible = false
-            }
-        }
-    }
-
-    private fun setupRecyclerView(): UserAdapter {
-        val recyclerView = findViewById<RecyclerView>(R.id.usersRecyclerView)
-        val adapter = UserAdapter()
-        recyclerView.adapter = adapter
-        return adapter
-    }
-
 }
