+++ app/src/main/java/com/daniillyubaev/ourawesomeapp/UserListFragment.kt
@@ -1,0 +1,61 @@
+package com.daniillyubaev.ourawesomeapp
+
+import android.os.Bundle
+import android.view.View
+import androidx.core.view.isVisible
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
+import by.kirich1409.viewbindingdelegate.viewBinding
+import com.daniillyubaev.ourawesomeapp.databinding.FragmentUserlistBinding
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
+
+class UserListFragment : Fragment(R.layout.fragment_userlist) {
+
+    private lateinit var viewModel: MainViewModel
+    private val viewBinding by viewBinding(FragmentUserlistBinding::bind)
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        viewModel = ViewModelProvider(this)[MainViewModel::class.java]
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setupRecyclerView()
+        lifecycleScope.launch {
+            lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                viewModel.viewState.collect { viewState ->
+                    renderViewState(viewState)
+                }
+            }
+        }
+    }
+
+    private fun setupRecyclerView(): UserAdapter {
+        val recyclerView = viewBinding.usersRecyclerView
+        val adapter = UserAdapter()
+        recyclerView.adapter = adapter
+        return adapter
+    }
+
+    private fun renderViewState(viewState: MainViewModel.ViewState) {
+        when (viewState) {
+            is MainViewModel.ViewState.Loading -> {
+                viewBinding.usersRecyclerView.isVisible = false
+                viewBinding.progressBar.isVisible = true
+            }
+            is MainViewModel.ViewState.Data -> {
+                viewBinding.usersRecyclerView.isVisible = true
+                (viewBinding.usersRecyclerView.adapter as UserAdapter).apply{
+                    userList = viewState.userList
+                    notifyDataSetChanged()
+                }
+                viewBinding.progressBar.isVisible = false
+            }
+        }
+    }
+}
